refactor(components): rename and update TextField structure
Moved and renamed TextField-related files to align with project conventions. Updated imports, typings, and definitions to improve consistency and maintainability. Simplified type usage by replacing redundant
