--- src/text-inputs/index.ts
@@ -1,3 +1,0 @@
-export * from './text-field.components';
-export * from './text-field.interface';
-export * from './text-field.style';
